+++ FuzzySharp/Delegates.cs
@@ -2,6 +2,6 @@
 
 namespace Raffinert.FuzzySharp;
 
-public delegate int Scorer(string input1, string input2);
-public delegate int CachedScorer(string input2);
+public delegate double Scorer(string input1, string input2);
+public delegate double CachedScorer(string input2);
 public delegate void Processor<T>(ref ReadOnlySpan<T> str) where T : IEquatable<T>;
