+++ FuzzySharp/Extractor/ExtractedResult.cs
@@ -3,18 +3,18 @@ using System.Collections.Generic;
 
 namespace Raffinert.FuzzySharp.Extractor;
 
-public class ExtractedResult<T>(T value, int score, int index) : IComparable<ExtractedResult<T>>
+public class ExtractedResult<T>(T value, double score, int index) : IComparable<ExtractedResult<T>>
 {
     public readonly T Value = value;
-    public readonly int Score = score;
+    public readonly double Score = score;
     public readonly int Index = index;
 
-    public ExtractedResult(T value, int score) : this(value, score, 0)
+    public ExtractedResult(T value, double score) : this(value, score, 0)
     { }
 
     public int CompareTo(ExtractedResult<T> other)
     {
-        return Comparer<int>.Default.Compare(this.Score, other.Score);
+        return Comparer<double>.Default.Compare(this.Score, other.Score);
     }
 
     public override string ToString()
