+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/Generic/StrategySensitiveScorerBase.cs
@@ -5,5 +5,5 @@ namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive.Generic;
 
 public abstract class StrategySensitiveScorerBase<T> : ScorerBase<T> where T : IEquatable<T>
 {
-    protected abstract Func<T[], T[], int> Scorer { get; }
+    protected abstract Func<T[], T[], double> Scorer { get; }
 }
