+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/Simple/CachedSimpleRatioScorerBase.cs
@@ -5,7 +5,7 @@ public abstract class CachedSimpleRatioScorerBase : ICachedRatioScorer
     protected abstract CachedScorer Scorer { get; }
     public abstract void Dispose();
 
-    public int Score(string input2)
+    public double Score(string input2)
     {
         return Scorer(input2);
     }
