+++ FuzzySharp/SimilarityRatio/Strategy/ICachedStrategy.cs
@@ -4,5 +4,5 @@ namespace Raffinert.FuzzySharp.SimilarityRatio.Strategy;
 
 public interface ICachedStrategy : IDisposable
 {
-    int Calculate(string input2);
+    double Calculate(string input2);
 }
